backend/requirements.txt
Django==2.2.27
pytz==2021.3
python-dotenv==0.19.2
sorl-thumbnail==12.7.0
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==21.1
fpdf==1.7.2
gunicorn==20.0.4
psycopg2-binary==2.8.6